Introducing `project` subcommand.
Instead of having to pass in connection details or rely on a fixed
set of environment variables, the project subcommand would save multiple
connetion details in ~/.livekit/cli-config.yaml.

The design of the CLI takes inspiration from kubectl, where all commands
are able to take in a `--project` flag and automatically use details from
that project.